Move `Basket` to a submodule
