+++ src/CurrencyFormatting/CurrencyFormatting.jl
@@ -1,5 +1,6 @@
 module CurrencyFormatting
 
+using ..Baskets
 using ..CurrenciesBase
 using ..CurrencyData
 using ..DeclarativeFormatting
